requirements.txt
torch --index-url https://download.pytorch.org/whl/cu124
torchvision --index-url https://download.pytorch.org/whl/cu124
torchaudio --index-url https://download.pytorch.org/whl/cu124

opencv-python
numpy
torchinfo
matplotlib
mlxtend
torchmetrics
tqdm
